Passing a pointer to dFord's RGB[3] array.
This will pass the pointer to the rgb[3] from your testColorHolder
function. I don’t have any LEDs to test the code on, but it compiles
according to the Arduino IDE so now the ball’s in your court ;). Will
see if I can help you out and explain it to you when you get a chance
to look at it.